requirements.txt
authy==2.2.3
phonenumbers==8.10.15
python-dotenv==0.10.3
django==2.0.7
psycopg2==2.7.7
djangorestframework
django-session-timeout
django-session-idle-timeout
django-crispy-forms
djangorestframework_simplejwt